+++ packages/bounds/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @solid-primitives/bounds
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [636b436]
+  - @solid-primitives/resize-observer@2.2.0
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [62e331d]
+  - @solid-primitives/event-listener@2.4.6
+  - @solid-primitives/utils@6.4.1
+  - @solid-primitives/resize-observer@2.1.6
+  - @solid-primitives/static-store@0.1.4
+
 ## 0.1.5
 
 ### Patch Changes