+++ packages/cursor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/cursor
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [62e331d]
+  - @solid-primitives/utils@6.4.1
+
 ## 0.1.3
 
 ### Patch Changes