+++ packages/destructure/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/destructure
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [62e331d]
+  - @solid-primitives/utils@6.4.1
+
 ## 0.2.3
 
 ### Patch Changes