+++ packages/intersection-observer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/intersection-observer
 
+## 2.2.5
+
+### Patch Changes
+
+- Updated dependencies [62e331d]
+  - @solid-primitives/utils@6.4.1
+
 ## 2.2.4
 
 ### Patch Changes