+++ packages/memo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @solid-primitives/memo
 
+## 1.5.1
+
+### Patch Changes
+
+- Updated dependencies [62e331d]
+  - @solid-primitives/utils@6.4.1
+  - @solid-primitives/scheduled@1.5.3
+
 ## 1.5.0
 
 ### Minor Changes