+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @solid-primitives/storage
 
+## 4.4.0
+
+### Minor Changes
+
+- 348974b: types and resource usage
+
+## 4.3.5
+
+### Patch Changes
+
+- Updated dependencies [62e331d]
+  - @solid-primitives/utils@6.4.1
+
 ## 4.3.4
 
 ### Patch Changes